requirements.txt
numpy
pandas
scikit-learn
joblib
pyserial
uvicorn[standard]
fastapi
bluezero